Create simpleTestingPipeline.py
Pipeline for running a test on two simple pruned CNNs. One is pruned with random unstructured pruning and the other one with magnitude based pruning L1 Unstrucured. 

First the unpruned model is evaluated on accuracy. Two instances from the class CNN are created and the unpruned model is loaded. Then both instances are pruned separately with the different pruning techniques and their initial accuracy is calculated. After that, cross-validation is performed on both pruned models assessing accuracy, f1 score, and training time on each fold of the cross-validation. Average and standard deviation of the results are calculated and stored into a txt file.

# TODO
Confusion matrices
Structured Pruning
Method implementation based on some paper
Some functions are still redundant.